move code to shell.py
